(1/2)Updated Bootanimatons for Kitkat! [TEST THIS]
All Credits go to crazy Ken(InkSlinger666) and his Son(Grim Gentleman)!

Notes: I'm not sure if I got the files in the proper spots, because of
this I've included the various zips and sizes for maintainers to sort
it out if i've got them wrong.

Still likely a work in progress, smooth some things out and invent a new
magic compression method to make them smaller but just as long and as
good looking as we currently have.

PS2: Alright, So i'm still not 100% sure on the sizes so i'm leaving the
 above parts. Once again a huge thanks again to Ken and Son for the
 initial images and the working concept. As well, thank you to Jason for
 the kick-ass work and expertize of boot animations shrinking the zips
 to our >10mb limit.

PS3&4&5: TO make the bootanimation look the best we possibly can, we're
 defining the choice directly by size of the screens in pixels. This may
 mean devices maintainers will have to fix the boot animation in device
 files. Please see the readme file that was added for more info.

PS6&7: Now we have all the sizes we could need!

PS8: It should have been obvious, but his is only one part of the
 updated. Now the title reflects that.

Change-Id: I527e00acd0ddb395cd696e4b23f065ec079434c2