Re-organize the Contributors list.
It's that of year again where we re-organize the contributors list. To those
who got demoted to contributors, it's because you haven't contribute that
much/nothing to the development (Coding) of PAC for a long time. And if in the
future, you still haven't contribute something to the development of PAC,
you'll get demoted to maintainers.
And for those who got promoted to contributors, congrats. You earned it!

From now on, it will be really difficult for someone to get promoted to
Developers. Do not ask us to promote you to Developers, we (Core Developers)
are the one who will decide if you earned it or not.

Please note for the people in Contributors category, if you're using the
Developers Banner, please take it off or switch it with something else.

Thank you!

Change-Id: I13cf09ef4983a9a07c939099d9cd9678531c3054
Signed-off-by: Lokesh Chamane <lokesh.c703@gmail.com>