Switch to my fork of android_device_samsung_smdk4412-common
My fork includes a revert of face detection for the camera. This fixes some quite common
crashes on S3 and possibly other devices. This is the only change compared to CM's repo.
We'll switch it back as soon as CM merges it (which they haven't for the last three months).
I'll keep my fork up to date with CM for the meantime.

Change-Id: Iac2776d43771cc7defdaab645582e6a67bf98c59