Insert "Sprint Update Menu" to Settings.APK
This might work, it compiled fine, but I doubt I've nailed down the integration of\
the actual PRL/Profile updater APKs. And even when I get them I'm not sure how
they'll look and if the changes here actually invoked the same concept from the
'toroplus' that I'm stealing the idea from.

Change-Id: I58093666b1902ea2d2d8dceec6401e83eb2358f9