build-pac.sh: Re-download prebuilts twice per month
The current script skips downloading the terminal apk if the folder exists,
but this means that the builds may be stuck with an outdated version of
the terminal application. This becomes especially annoying if updating
nightlies as the Play Store will have to re-download the latest version
every time after flashing a build. This patch forces a re-download of the
prebuilts on the 1st and 15th of every month.

Change-Id: I01b4c5b19b022ed03c5029be248718d4426a3e3c
Signed-off-by: Tianrui Guo <tianrui@pac-rom.com>