Falcon depedencies: re-switch to CyanogenMod repos
It seems that intervigilium is the new maintainer of this falcon device
in CyanogenMod and he already made a lot of changes.
When looking at Team Gummy's falcon repos, there are not so many new
commits (there was no commit this week).

If we want to use Team Gummy's repos, we can't just use one repo
(android_kernel_motorola_msm8226) but most of them in falcon.dependencies
file.

But because Moto G seems to be still maintained in CyanogenMod repos and
because it no longer compiles when using android_kernel_motorola_msm8226
repo from Team Gummy, I guess it's better to switch back to CyanogenMod
repos only.

Note: android_device_motorola_qcom-common repo is no longer needed.

Change-Id: I176dc4a4536e1e97a1fa9af72e261612e7224860